@@ -3,8 +3,7 @@ package com.algaworks.algafood.api.controller;
 import java.util.List;
 import java.util.Map;
 
-import javax.validation.Valid;
-
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
