@@ -1,7 +1,6 @@
 package com.algaworks.algafood.api.model.input;
 
-import javax.validation.constraints.NotBlank;
-
+import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
 
