@@ -1,8 +1,7 @@
 package com.algaworks.algafood.api.model.input;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
 
