@@ -4,14 +4,14 @@ import java.time.OffsetDateTime;
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.ManyToMany;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinTable;
+import jakarta.persistence.ManyToMany;
 
 import org.hibernate.annotations.CreationTimestamp;
 
@@ -46,20 +46,16 @@ public class Usuario {
 			inverseJoinColumns = @JoinColumn(name = "grupo_id"))
 	private Set<Grupo> grupos = new HashSet<>();
 	
-	public boolean senhaCoincideCom(String senha) {
-		return getSenha().equals(senha);
-	}
-	
-	public boolean senhaNaoCoincideCom(String senha) {
-		return !senhaCoincideCom(senha);
-	}
-	
 	public boolean removerGrupo(Grupo grupo) {
 		return getGrupos().remove(grupo);
 	}
 	
 	public boolean adicionarGrupo(Grupo grupo) {
 		return getGrupos().add(grupo);
+	}
+	
+	public boolean isNovo() {
+		return getId() == null;
 	}
 	
 }
