@@ -1,18 +1,14 @@
 package com.algaworks.algafood.infrastructure.service.email;
 
-import javax.mail.MessagingException;
-import javax.mail.internet.MimeMessage;
+import jakarta.mail.MessagingException;
+import jakarta.mail.internet.MimeMessage;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.ui.freemarker.FreeMarkerTemplateUtils;
 
 import com.algaworks.algafood.core.email.EmailProperties;
 import com.algaworks.algafood.domain.service.EnvioEmailService;
-
-import freemarker.template.Configuration;
-import freemarker.template.Template;
 
 public class SmtpEnvioEmailService implements EnvioEmailService {
 
@@ -23,7 +19,7 @@ public class SmtpEnvioEmailService implements EnvioEmailService {
 	private EmailProperties emailProperties;
 	
 	@Autowired
-	private Configuration freemarkerConfig;
+	private ProcessadorEmailTemplate processadorEmailTemplate;
 	
 	@Override
 	public void enviar(Mensagem mensagem) {
@@ -37,7 +33,7 @@ public class SmtpEnvioEmailService implements EnvioEmailService {
 	}
 	
 	protected MimeMessage criarMimeMessage(Mensagem mensagem) throws MessagingException {
-		String corpo = processarTemplate(mensagem);
+		String corpo = processadorEmailTemplate.processarTemplate(mensagem);
 		
 		MimeMessage mimeMessage = mailSender.createMimeMessage();
 		
@@ -50,15 +46,4 @@ public class SmtpEnvioEmailService implements EnvioEmailService {
 		return mimeMessage;
 	}
 	
-	protected String processarTemplate(Mensagem mensagem) {
-		try {
-			Template template = freemarkerConfig.getTemplate(mensagem.getCorpo());
-			
-			return FreeMarkerTemplateUtils.processTemplateIntoString(
-					template, mensagem.getVariaveis());
-		} catch (Exception e) {
-			throw new EmailException("Não foi possível montar o template do e-mail", e);
-		}
-	}
-
 }
