@@ -6,6 +6,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
+import com.algaworks.algafood.core.io.Base64ProtocolResolver;
 import com.algaworks.algafood.infrastructure.repository.CustomJpaRepositoryImpl;
 
 @SpringBootApplication
@@ -15,7 +16,9 @@ public class AlgafoodApiApplication {
 	public static void main(String[] args) {
 		TimeZone.setDefault(TimeZone.getTimeZone("UTC"));
 		
-		SpringApplication.run(AlgafoodApiApplication.class, args);
+		var app = new SpringApplication(AlgafoodApiApplication.class);
+		app.addListeners(new Base64ProtocolResolver());
+		app.run(args);
 	}
 
 }
