@@ -4,8 +4,8 @@ import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import com.algaworks.algafood.api.model.EnderecoModel;
-import com.algaworks.algafood.api.model.input.ItemPedidoInput;
+import com.algaworks.algafood.api.v1.model.EnderecoModel;
+import com.algaworks.algafood.api.v1.model.input.ItemPedidoInput;
 import com.algaworks.algafood.domain.model.Endereco;
 import com.algaworks.algafood.domain.model.ItemPedido;
 
@@ -19,7 +19,7 @@ public class ModelMapperConfig {
 //		modelMapper.createTypeMap(Restaurante.class, RestauranteModel.class)
 //			.addMapping(Restaurante::getTaxaFrete, RestauranteModel::setPrecoFrete);
 		
-		/*modelMapper.createTypeMap(ItemPedidoInput.class, ItemPedido.class)
+		modelMapper.createTypeMap(ItemPedidoInput.class, ItemPedido.class)
 			.addMappings(mapper -> mapper.skip(ItemPedido::setId));
 		
 		var enderecoToEnderecoModelTypeMap = modelMapper.createTypeMap(
@@ -27,7 +27,7 @@ public class ModelMapperConfig {
 		
 		enderecoToEnderecoModelTypeMap.<String>addMapping(
 				enderecoSrc -> enderecoSrc.getCidade().getEstado().getNome(),
-				(enderecoModelDest, value) -> enderecoModelDest.getCidade().setEstado(value));*/
+				(enderecoModelDest, value) -> enderecoModelDest.getCidade().setEstado(value));
 		
 		return modelMapper;
 	}
