@@ -1,18 +1,18 @@
 package com.algaworks.algafood.domain.filter;
 
-import java.time.OffsetDateTime;
-
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.format.annotation.DateTimeFormat.ISO;
 
-import lombok.Getter;
-import lombok.Setter;
+import java.time.OffsetDateTime;
 
 @Setter
 @Getter
 public class PedidoFilter {
 
 	private Long clienteId;
+	
 	private Long restauranteId;
 	
 	@DateTimeFormat(iso = ISO.DATE_TIME)
